+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### [0.8.1](Unreleased)
 
 #### Fixed
+- fix: add o2o field does not create constraint when migrating. (#396)
 - fix: intermediate table for m2m relation not created. (#394)
 - Migrate add m2m field with custom through generate duplicated table. (#393)
 - Migrate drop the wrong m2m field when model have multi m2m fields. (#376)
@@ -15,10 +16,12 @@
 - Fix configuration file reading error when containing Chinese characters. (#286)
 - sqlite: failed to create/drop index. (#302)
 - PostgreSQL: Cannot drop constraint after deleting or rename FK on a model. (#378)
+- Fix create/drop indexes in every migration. (#377)
 - Sort m2m fields before comparing them with diff. (#271)
 
 #### Changed
 - Allow run `aerich init-db` with empty migration directories instead of abort with warnings. (#286)
+- Add version constraint(>=0.21) for tortoise-orm. (#388)
 - Move `tomlkit` to optional and support `pip install aerich[toml]`. (#392)
 
 ### [0.8.0](../../releases/tag/v0.8.0) - 2024-12-04